Refactor go-script packages to use the new import functionality within go-script
