+++ script/init.go
@@ -1,19 +1,11 @@
 package script
 
 import (
+	_ "github.com/peter-mount/go-anim/script/colour"
 	_ "github.com/peter-mount/go-anim/script/exif"
-	_ "github.com/peter-mount/go-anim/script/io"
-	"github.com/peter-mount/go-script/packages"
+	_ "github.com/peter-mount/go-anim/script/graph"
+	_ "github.com/peter-mount/go-anim/script/image"
+	_ "github.com/peter-mount/go-anim/script/mapper"
+	_ "github.com/peter-mount/go-anim/script/render"
+	_ "github.com/peter-mount/go-anim/script/util"
 )
-
-func init() {
-	packages.Register("animGraphic", &Graph{})
-	packages.Register("animUtil", &AnimUtil{})
-	packages.Register("colour", &Colour{})
-	packages.Register("draw2dimg", &Draw2Dimg{})
-	packages.Register("image", newImage())
-
-	packages.Register("graphFilter", Filter{})
-	packages.Register("graphMapper", Mapper{})
-
-}
