+++ script/render/frames.go
@@ -1,4 +1,4 @@
-package io
+package render
 
 import (
 	"fmt"
